Removed doc and sources from source files
I moved the documentation to the new doc/ directory and removed the
line with all the sources I used for writing the source files. That
just became tedious and if it becomes tedious at such an early stage
then it'll be a real pain later on. I have all the sources in
SOURCES and that is completely sufficient.