Added memory detection functionality
The memory detection so far only uses ah=0xe820, int=0x15 and even that
is still very basic, but at least we got something implemented now. I'm still
unsure about how we're going to utilize the memory detection, but there is
a use case for it for sure.
Still many TODOs there, so work on those and finish the implementation!