Add stub for the header file for the boot paramters
The boot system passes boot parameters to the kernel. The boot system
is currently written in assembler and the kernel will be written in C,
so, to help maintenance, I want a clean interface between both parts.
That means, macros for the structure containing the parameters and
those macros expand to either assembly or C conditionally.