Add kernel loading
The kernel is loaded into memory now. What is missing is the jump into
Protected Mode and into the kernel. I moved the floppy constants
to the boot m4 file because the kernel chs scheme is depending on the chs
values of the bootloader because the kernel is supposed to follow the bootloader
directly on disk, so i decided that they belong together and the boot.m4 file
is the best-suited location for that.