Fix GDTR register loading
While in Protected Mode a linear address is passed without segment
selector, in Real Mode, a segment is still used as is done for
every address. Therefore, a simple `lgdt offset` is sufficient,
no physical address needs to be computed.
I also changed GDTR's content, so the limit is at the lower
address as assumed by `lgdt`.