Remove debugging from makefile (temporarily)
Some debugging stuff was in the topmost makefile, but the
initialization of bochs is so trivial that I don't need it ATM.
I removed QEMU support for the boot code because it isn't capable
of debugging 16-bit code anyway. Later, when the kernel itself will
be written (32-bit), QEMU support will return for the kernel.
But we're not there yet.