Convert i8042 helper routines to macros
In the former implementation of the Intel 8042 driver, I used a lot
of `call's for even very simple things. I converted a lot of them
to macros now. I also renamed the macros for the command values to
avoid collisions with the new macros and renamed some helper macros
to more descriptive names.