Add header file for parameters between primary and secondary bootloader
Adds a header file for a clean interface for passing parameters
from the primary to the secondary bootloader.
I'm committing this at home and just because it somehow isn't part
of the repository. At the boarding school, I already have it saved
as a file. Maybe I forgot to track and commit it, but I don't
remember it to be shown as untracked when doing `git status`...
anyway, I just add it here, that fixes the problem.