Add cursor movement
For every line, the cursor is moved now. Moving the cursor
at the very end of all printing would be faster, but I
have a feeling moving it for every line is better. And
we don't care about speed anyway, only size.
I also changed the IOS bit from the CRTC Miscellaneous
Output register to `1'. `0' is for monochrome, `1' for
color, I think.