Add disk error treatment
I added a proper error message for when a disk operation fails. The
thing is that we sort of have a chicken-and-egg problem here: code
and data for disk operations with full error messages for every
single error are on disk, so they must be fetched from disk by the
very disk operations that can fail and which need proper treatment.
And I cannot store all the full error messages in the primary
bootloader because then it gets too large. Eventually, I tackled this
problem by telling the user the error code through a message and let
him figure out the corresponding error message on his own via the
Internet or some other means.