Clean descriptor table initialization up
I basically cleaned up the table initialization by putting
everything into separate files and involving C. The
descriptors are no longer numeric constants but consist
of macros, which are way more readable. So far, I've done
so only with the GDT, but the same improvements for the IDT
are under way.