Add memory detection using ah=0xe820,int 0x15
Some things were seriously wrong:
1. The macro for a memory map entry was wrong. 20 = 0x14,
not 0x16!

2.  Some registers were being initialized for every iteration,
even though they are not touched by BIOS. I know, BIOS can
can be faulty, but we assume this is not the case here.
If we did the opposite, I would have to build a proper
"BIOS-sandbox," which controls all BIOS calls.

3.  `%ds' points to the boot system already, `%es' can be used
for other purposes, though, so the access uses `%es' now.

4. The incrementation of the entry index was uncommented. It was
commented before for testing.