MAJOR REVAMP
So... where should I start? There is a large number of things I have changed
to improve the code and make it work nicely. The main things are:

Changed memory layout

The whole memory layout of the boot system + kernel has been changed to
make them work smoothly together. You can read up about it in the doc/
directory.

cpp as the new main preprocessor

While GNU m4 is still supported, cpp is now the preprocessor of choice. I still
consider GNU m4 to be the one more suitable for this project IN THEORY, but
unfortunately, it has turned out to be kind of a disappointment, I must say.
The GNU m4 maintainers rarely commit anything at all, let alone adding new
features. What I miss most is a switch to remove the comments. ATM, GNU m4
parses them and if, like, one of the quotation marks, which default to `
and ', occur there, the parser might get confused. There is such a switch
implemented in the GNU m4 1.6 (?) branch, but it has not been released yet.
I seriously contemplated chiming in as a dev myself and submitting some patches,
but the maintainers don't seem to be too receptive to that. I'm not blaming
them, I'm sure they are really busy (no irony, seriously), but it is a pity
nevertheless. And getting familiar with the code base also requires some
serious effort... I don't want that right now. Maybe some time in the future.
All in all, m4 is a better fit as a GENERAL preprocessor, but syntactical
constraints + lacking dev efforts have made me opt for cpp. Bummer.

Make the sector numbering flexible in the root makefile

This is minor compared to the other two things, but it counts as a separate
change. The hard-coded values in the makefile have already caused two
nasty bugs, which were hard to track down. The makefile is preprocessed now
and macros are used instead of the literals. Building is done using the
`build' script in the root directory now. I hope it'll turn out to be
maintainable.

I did not split this up into separate commits because, as a matter of fact,
this is the first commit after a while and the changes are all intertwined.
I fucked up a little because I developed absolutely ignoring git. I did
set up a new branch, but from then on, I pretty much did development like
I used to do back in the day: basically a whole mess. I'm still not really
acquainted with git, but commits really help organizing your work and your
mind.

Note: this was intended to be a rewrite of the boot system in C. I had some
problems with 32-bit addressing in Unreal Mode caused by GNU as, though,
and after a brief discussion with someone on SO, I reverted to assembly.