Improve support for bss symbols
I was looking for a way to define symbols and reserve space for them
IN MEMORY, not in the image. I thought bss symbols need special runtime
support, but I was mistaken, so I am using the bss section for data
such as the kernel page directory now: this directory is not
statically initialized but dynamically and storing it in the image
would be a huge waste of space.
There is just one problem: the kernel linker script has a "kernel"
memory area, which makes the linker raise an error should the kernel
get so large that the number of sectors loaded does not suffice to
bring the whole kernel into memory. This is pretty handy but
unfortunately, it also accounts for the space allocated in the bss
section. The bss section doesn't occupy any space in the actual image,
though! It's just symbols with associated lengths, whose memory is
initialized at runtime. So, if the maximum kernel length is 16 KiB and
the kernel image is 10 KiB, the image fits adequately. However, if the
bss section contains symbols reserving space larger than 6 KiB, the
linker speaks up, even though the bss section does not take up any space
in the image and is not loaded into memory.
That is pretty bad. My solution so far is to load the actual data, which
really is in the image, and the bss section into separate linker
memory areas (see MEMORY). This solves the current problem, but is a
little space-inefficient because, if the kernel does not align to a
sector-boundary, some space is wasted: number of sectors * sector size
- kernel size, to be precise. I don't know how to overcome this yet,
though. The amount of wasted memory is little, but it should be
avoidable in theory. And accepting inefficiencies because of
constraints imposed by the tools I use is something I absolutely
detest. Maybe it's even me who is using the toolchain wrongly, but
either way, the linker should not hinder me from enhancing this.